feat: improve webpack config
