+++ client/index.html
@@ -19,8 +19,8 @@
 
 <body style="margin: 0px; background: #FAFAFA;">
   <div id="lumi"></div>
-  <script src="/vendor.js"></script>
-  <script src="/client.js"></script>
+  <script defer src="/vendor.js"></script>
+  <script defer src="/client.js"></script>
 </body>
 
 </html>
