+++ package.json
@@ -11,9 +11,9 @@
     "test": "NODE_ENV=test ./node_modules/.bin/mocha --watch --watch-extensions tsx,ts,test.ts --require ts-node/register --recursive ./**/*.test.ts",
     "clean": "rm -rf build && rm -rf dist",
     "uninstall": "npm run clean && rm -rf node_modules",
-    "dev": "NODE_ENV=development ./node_modules/.bin/webpack-dev-server --config './webpack.js' --progress --colors",
+    "dev": "NODE_ENV=development ./node_modules/.bin/webpack-dev-server --config './config/webpack.config.dev.js' --progress --colors --hot --inline",
     "build:server": "./node_modules/.bin/tsc server/boot.ts --sourceMap --inlineSources --module commonjs --target es5 --outDir build/",
-    "build:client": "NODE_ENV=production webpack -p --config './webpack.js' --progress --colors",
+    "build:client": "NODE_ENV=production webpack -p --config './config/webpack.config.prod.js' --progress --colors",
     "build:docker": "npm run clean && npm run build:client && docker build -t jpschellenberg/lumi_user:develop .",
     "build": "npm run clean && npm run build:client && npm run build:server",
     "link": "node -e \"var s='../client',d='node_modules/client',fs=require('fs');fs.exists(d,function(e){e||fs.symlinkSync(s,d,'dir')});\" && node -e \"var s='../server',d='node_modules/server',fs=require('fs');fs.exists(d,function(e){e||fs.symlinkSync(s,d,'dir')});\" && node -e \"var s='../lib',d='node_modules/lib',fs=require('fs');fs.exists(d,function(e){e||fs.symlinkSync(s,d,'dir')});\"",
